check cross-ns reference grant
